# docs/contracts/openapi/platform-tenant-service.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://127.0.0.1:4110","description":"Generated server url"}],"paths":{"/admin/tenant/currency":{"get":{"tags":["tenant-currency-controller"],"operationId":"get","parameters":[{"name":"tenantId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CurrencyView"}}}}}},"put":{"tags":["tenant-currency-controller"],"operationId":"update","parameters":[{"name":"tenantId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCurrencyRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CurrencyView"}}}}}}},"/admin/tenant/config":{"get":{"tags":["tenant-config-controller"],"operationId":"get_1","parameters":[{"name":"tenantId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WalletTokenConfig"}}}}}},"put":{"tags":["tenant-config-controller"],"operationId":"update_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WalletTokenConfig"}}}}}}},"/internal/pricing-plans":{"get":{"tags":["internal-pricing-plan-controller"],"operationId":"list","parameters":[{"name":"storeId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PricingPlanPo"}}}}}}},"post":{"tags":["internal-pricing-plan-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePricingRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PricingPlanPo"}}}}}}},"/internal/iam/consumer/{appId}/authorizations":{"post":{"tags":["internal-iam-controller"],"operationId":"grantConsumerAuthorization","parameters":[{"name":"appId","in":"path","required":true,"schema":{"type":"string"}},{"name":"accountId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"scope","in":"query","required":false,"schema":{"type":"string","default":"profile.basic"}}],"responses":{"200":{"description":"OK"}}}},"/admin/pricing-plans":{"post":{"tags":["pricing-plan-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePricingRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PricingPlanPo"}}}}}}},"/admin/platform/tenants":{"get":{"tags":["tenant-controller"],"operationId":"list_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TenantPo"}}}}}}},"post":{"tags":["tenant-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTenantRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TenantPo"}}}}}}},"/admin/iam/user-roles":{"post":{"tags":["iam-controller"],"operationId":"assignUserRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignUserRoleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRolePo"}}}}}}},"/admin/iam/roles":{"get":{"tags":["iam-controller"],"operationId":"listRoles","parameters":[{"name":"tenantId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RolePo"}}}}}}},"post":{"tags":["iam-controller"],"operationId":"createRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RolePo"}}}}}}},"/admin/iam/roles/{roleId}/permissions":{"get":{"tags":["iam-controller"],"operationId":"listRolePermissions","parameters":[{"name":"roleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}},"post":{"tags":["iam-controller"],"operationId":"assignPermissions","parameters":[{"name":"roleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignPermissionsRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/admin/iam/roles/{roleId}/permissions/toggle":{"post":{"tags":["iam-controller"],"operationId":"togglePermission","parameters":[{"name":"roleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TogglePermissionRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/admin/iam/approvals":{"get":{"tags":["iam-approval-controller"],"operationId":"list_2","parameters":[{"name":"actionType","in":"query","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IamApproval"}}}}}}},"post":{"tags":["iam-approval-controller"],"operationId":"submit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IamApproval"}}}}}}},"/admin/iam/approvals/{id}/reject":{"post":{"tags":["iam-approval-controller"],"operationId":"reject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IamApproval"}}}}}}},"/admin/iam/approvals/{id}/approve":{"post":{"tags":["iam-approval-controller"],"operationId":"approve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IamApproval"}}}}}}},"/internal/iam/tenant-config/{tenantId}/currency":{"get":{"tags":["internal-tenant-config-controller"],"operationId":"tenantCurrency","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InternalTenantCurrency"}}}}}}},"/internal/iam/consumer/{appId}/permissions":{"get":{"tags":["internal-iam-controller"],"operationId":"consumerPermissions","parameters":[{"name":"appId","in":"path","required":true,"schema":{"type":"string"}},{"name":"accountId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PermissionSnapshot"}}}}}}},"/internal/iam/consumer/{appId}/context":{"get":{"tags":["internal-iam-controller"],"operationId":"consumerContext","parameters":[{"name":"appId","in":"path","required":true,"schema":{"type":"string"}},{"name":"accountId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountContext"}}}}}}},"/internal/iam/account/{accountId}/role-bindings":{"get":{"tags":["internal-iam-controller"],"operationId":"roleBindings","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoleBinding"}}}}}}}},"/internal/iam/account/{accountId}/permissions":{"get":{"tags":["internal-iam-controller"],"operationId":"permissions","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"tenantId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"organizationId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"storeId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"refresh","in":"query","required":false,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PermissionSnapshot"}}}}}}},"/internal/iam/account/{accountId}/contexts":{"get":{"tags":["internal-iam-controller"],"operationId":"contexts","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountContext"}}}}}}}},"/admin/tenant/stores":{"get":{"tags":["store-controller"],"operationId":"list_3","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StorePo"}}}}}}}},"/admin/iam/permissions":{"get":{"tags":["iam-controller"],"operationId":"listPermissions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PermissionPo"}}}}}}}},"/admin/iam/masking":{"get":{"tags":["iam-controller"],"operationId":"masking","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MaskingRoleView"}}}}}}}},"/admin/iam/approvals/{id}":{"get":{"tags":["iam-approval-controller"],"operationId":"get_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IamApproval"}}}}}}},"/admin/iam/accounts/{accountId}/user-roles":{"get":{"tags":["iam-controller"],"operationId":"listUserRoleIds","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"integer","format":"int64"}}}}}}}},"/admin/iam/user-roles/{userRoleId}":{"delete":{"tags":["iam-controller"],"operationId":"revokeUserRole","parameters":[{"name":"userRoleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UpdateCurrencyRequest":{"type":"object","properties":{"currencyCode":{"type":"string"},"migrateBalances":{"type":"boolean"}}},"CurrencyView":{"type":"object","properties":{"currencyCode":{"type":"string"},"symbol":{"type":"string"},"minorUnitDigits":{"type":"integer","format":"int32"},"supported":{"type":"array","items":{"$ref":"#/components/schemas/SupportedCurrency"}}}},"SupportedCurrency":{"type":"object","properties":{"code":{"type":"string"},"symbol":{"type":"string"},"label":{"type":"string"}}},"UpdateRequest":{"type":"object","properties":{"tenantId":{"type":"integer","format":"int64"},"brandName":{"type":"string"},"ratio":{"type":"integer","format":"int64"}}},"WalletTokenConfig":{"type":"object","properties":{"brandName":{"type":"string"},"ratio":{"type":"integer","format":"int64"}}},"CreatePricingRequest":{"type":"object","properties":{"tenantId":{"type":"integer","format":"int64"},"storeId":{"type":"integer","format":"int64"},"resourceType":{"type":"string"},"billingUnit":{"type":"string"},"incrementMinutes":{"type":"integer","format":"int32"},"roundingDirection":{"type":"string"},"pricePerUnit":{"type":"integer","format":"int64"},"defaultSessionMinutes":{"type":"integer","format":"int32"},"overtimeRate":{"type":"number"}}},"PricingPlanPo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"tenantId":{"type":"integer","format":"int64"},"storeId":{"type":"integer","format":"int64"},"resourceType":{"type":"string"},"billingUnit":{"type":"string"},"incrementMinutes":{"type":"integer","format":"int32"},"roundingDirection":{"type":"string"},"pricePerUnit":{"type":"integer","format":"int64"},"defaultSessionMinutes":{"type":"integer","format":"int32"},"overtimeRate":{"type":"number"},"status":{"type":"string"},"createdBy":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedBy":{"type":"integer","format":"int64"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateTenantRequest":{"type":"object","properties":{"tenantCode":{"type":"string"},"name":{"type":"string"},"defaultLocale":{"type":"string"},"defaultTimezone":{"type":"string"}}},"TenantPo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"tenantCode":{"type":"string"},"name":{"type":"string"},"status":{"type":"string"},"defaultLocale":{"type":"string"},"defaultTimezone":{"type":"string"},"createdBy":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedBy":{"type":"integer","format":"int64"},"updatedAt":{"type":"string","format":"date-time"}}},"AssignUserRoleRequest":{"type":"object","properties":{"accountId":{"type":"integer","format":"int64"},"tenantId":{"type":"integer","format":"int64"},"organizationId":{"type":"integer","format":"int64"},"storeId":{"type":"integer","format":"int64"},"roleId":{"type":"integer","format":"int64"},"scopeType":{"type":"string"}}},"UserRolePo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"accountId":{"type":"integer","format":"int64"},"tenantId":{"type":"integer","format":"int64"},"organizationId":{"type":"integer","format":"int64"},"storeId":{"type":"integer","format":"int64"},"roleId":{"type":"integer","format":"int64"},"scopeType":{"type":"string"},"effectiveFrom":{"type":"string","format":"date-time"},"effectiveTo":{"type":"string","format":"date-time"},"status":{"type":"string"},"authorizationVersion":{"type":"integer","format":"int32"},"createdBy":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedBy":{"type":"integer","format":"int64"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateRoleRequest":{"type":"object","properties":{"tenantId":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"roleType":{"type":"string"}}},"RolePo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"tenantId":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"roleType":{"type":"string"},"copyFromRoleId":{"type":"integer","format":"int64"},"status":{"type":"string"},"createdBy":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedBy":{"type":"integer","format":"int64"},"updatedAt":{"type":"string","format":"date-time"}}},"AssignPermissionsRequest":{"type":"object","properties":{"permissionIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"TogglePermissionRequest":{"type":"object","properties":{"permissionCode":{"type":"string"},"granted":{"type":"boolean"}}},"SubmitRequest":{"type":"object","properties":{"tenantId":{"type":"integer","format":"int64"},"actionType":{"type":"string"},"resourceType":{"type":"string"},"resourceId":{"type":"string"},"operatorId":{"type":"integer","format":"int64"},"detailJson":{"type":"string"},"idempotencyKey":{"type":"string"}}},"IamApproval":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"tenantId":{"type":"integer","format":"int64"},"actionType":{"type":"string"},"resourceType":{"type":"string"},"resourceId":{"type":"string"},"operatorId":{"type":"integer","format":"int64"},"detailJson":{"type":"string"},"status":{"type":"string","enum":["PENDING","APPROVED","REJECTED"]},"idempotencyKey":{"type":"string"},"approverId":{"type":"integer","format":"int64"},"reviewComment":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"reviewedAt":{"type":"string","format":"date-time"}}},"ReviewRequest":{"type":"object","properties":{"approverId":{"type":"integer","format":"int64"},"comment":{"type":"string"}}},"InternalTenantCurrency":{"type":"object","properties":{"tenantId":{"type":"integer","format":"int64"},"currencyCode":{"type":"string"}}},"PermissionSnapshot":{"type":"object","properties":{"accountId":{"type":"integer","format":"int64"},"tenantId":{"type":"integer","format":"int64"},"organizationId":{"type":"integer","format":"int64"},"storeId":{"type":"integer","format":"int64"},"authorizationVersion":{"type":"integer","format":"int32"},"permissions":{"type":"array","items":{"type":"string"}}}},"AccountContext":{"type":"object","properties":{"contextId":{"type":"string"},"tenantId":{"type":"integer","format":"int64"},"tenantName":{"type":"string"},"organizationId":{"type":"integer","format":"int64"},"organizationName":{"type":"string"},"storeId":{"type":"integer","format":"int64"},"storeName":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"scopeType":{"type":"string"}}},"RoleBinding":{"type":"object","properties":{"roleId":{"type":"integer","format":"int64"},"tenantId":{"type":"integer","format":"int64"},"organizationId":{"type":"integer","format":"int64"},"storeId":{"type":"integer","format":"int64"},"roleCode":{"type":"string"},"roleName":{"type":"string"},"scopeType":{"type":"string"},"tenantName":{"type":"string"},"storeName":{"type":"string"},"status":{"type":"string"}}},"StorePo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"tenantId":{"type":"integer","format":"int64"},"organizationId":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"businessType":{"type":"string"},"countryCode":{"type":"string"},"regionCode":{"type":"string"},"timezone":{"type":"string"},"defaultCurrency":{"type":"string"},"locale":{"type":"string"},"taxProfileId":{"type":"integer","format":"int64"},"businessDayCutoff":{"type":"string"},"status":{"type":"string"},"createdBy":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedBy":{"type":"integer","format":"int64"},"updatedAt":{"type":"string","format":"date-time"}}},"PermissionPo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"module":{"type":"string"},"resource":{"type":"string"},"action":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"createdBy":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedBy":{"type":"integer","format":"int64"},"updatedAt":{"type":"string","format":"date-time"}}},"MaskingRoleView":{"type":"object","properties":{"roleId":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"piiView":{"type":"boolean"}}}}}}